Added a null check to SteamGameRegistryFinder
On linux and mono, the Steam registry lookup would return null which would crash the process.

```
[Nitrox] E: System.ArgumentNullException: Value cannot be null.
Parameter name: path1
  at System.IO.Path.Combine (System.String path1, System.String path2) [0x00003] in <3d1438d9da1e4fe0b915f807d1ba56bd>:0
  at NitroxModel.Discovery.InstallationFinders.SteamGameRegistryFinder.FindGame (System.Collections.Generic.List`1[T] errors) [0x0000f] in <484753777b4b4c1ea501fed6c0b21452>:0
```

This proposal ensures that this does not happen and that in the case that the Steam registry can not be found for whatever reason, the application continues to look for the game files from other sources.

This is not yet tested. I will verify this as soon as I have the chance.