Damage Syncing Complete
Despite a minor bug, Cyclops damage is ready to go.

The health state of the Cyclops is fully synced with everyone, including damage points and fires, which covers everything from engine overheats to collision damage.

Now to work on a couple of changes that will make the syncing feel a bit smoother.