Final Refactor
Finished final stage of refactoring.

Flattened SerializableDamageInfo, SerializableFireNode, and SerializableRoomFire to DamageInfoData and FireData to better fit the design pattern. DamageInfoData is separate from Cyclops, but used by the damage processor.

Separated out Fires from the Cyclops as much as possible. Cyclops fire creation seems to be the only way to create new fires. If a different method is found, it should be added to Fires.Create().

CyclopsDamagePoints remain tied to Cyclops.