Persistent vehicle batteries (#754)
* StorageSlot synchronization for all EnergyMixin objects.
There is one (minor) error, that needs to be addressed: When joining the server, all vehicles will be created with default Batteries. These will be send to other players (but cannot be processed, as guids are not synched to this time)
Need to switch off sending, if in starting process.

* StorageSlot synchronization for all EnergyMixin objects.
There is one (minor) error, that needs to be addressed: When joining the server, all vehicles will be created with default Batteries. These will be send to other players (but cannot be processed, as guids are not synched to this time)
Need to switch off sending, if in starting process.

* Persistent save for EnergyMixin storage save (Vehicle batteries)

* Repaired Constructor not synching. Now vehicles will be build on all clients.
Silent battery add mode for initial loading.
Also sorting vehicles before creating them to (hopefully) prevent problems with docking in cyclops

* Changes for review

* Renamed SendBatterySlots to a more fitting name

* Rework of the initial battery spawn for vehicles.
Now all vehicles will not spawn with initial batteries (all other items still will)
and initial batteries will be set later when synchronized.

* debug message removed

* removed dead code

* Remove batteries from server if vehicle is destroyed