Unity 2019 Upgrade (#864)
* Unity 2019 Upgrade

Updates the projects to .Net Framework 4.0 and removes references to libraries that no longer exist.

* Update PAXTerrainController_LoadAsync_Patch.cs

Added menu fix

* Update BaseDeconstructable_Deconstruct_Patch.cs

Updated this line to use ConstructableBase. This allows Nitrox to host games again after the Unity 2019 Update, but there seems to be an issue with serializing the data as nothing appears to be saved anymore.