Compile entire project & run server on Linux (#995)
* NitroxModel: Fix capitalization of InitialPDAData.cs

Capitailization with the vsproj file has to match on case-sensitive file
systems/os-es.

* NitroxModel: Fix capitalization of Packets/GameModeChanged.cs

* NitroxPatcher: Fix capitalization of uGUI_FeedbackCollector_IsEnabled_Patch

* Fix conditional escapes for xbuild bug

https://stackoverflow.com/a/29237754/2844473
https://xamarin.github.io/bugzilla-archives/19/19678/bug.html

* NitroxServer: Clean up confusing Visual Studio library vs executable

VS doesn't seem able to keep these xml files clean and understandable. A
followup <PropertyGroup> without conditions sets the OutputType to
Library again.

* NitroxServer-Subnautica: Fix Resources/ capitalization

AssetTools-NET is said to handle this properly but it doesn't, so
:shrug:

* NitroxServer-Subnautica: Prevent crashing in QuickEdit() on Linux

* NitroxLauncher: Prevent building on non-Windows platforms

Mono doesn't support WPF

* NitroxTest: Disable compilation on non-Windows OSes

Tests reference Microsoft.VisualStudio.TestTools.UnitTesting which is
not available under Mono.

Rather than attempting to "rewrite" all tests to use a different
framework, disable it for now until someone finds the time and courage
to do so.

* NitroxModel-Subnautica: Normalize inclusion of LitJson

Fixes: b73d55d085e0feca47caabf4a55175b6c7749208

* Revert "NitroxModel-Subnautica: Normalize inclusion of LitJson"

This reverts commit 8ed0fef618b75b2e57cbb168809a2aab524819da.

* NS-Subnautica: SubnauticaServerProtoBufSerializer capitalization

Normalize filename capitalization to match the class name, just like
ServerProtoBufSerializer.cs.

* NitroxServer: csproj: Fix capitalization of ServerProtoBufSerializer

This file has been recapitalized to match the classname, but the csproj
was conveniently forgotten about.

* NitroxServer-Subnautica: Set up logging before doing anything else

Logging is used in Main() and ConfigureConsoleWindow() before the logger
is set up, resulting in a nice NullReferenceException. Reorder the code
to ensure we can use it as early as possible without having any sort of
CultureInfo issues with formatting.