Started work on improving UI/UX in the World Settings pane and fixed another bug
- Replaced the "Start Server" button in the world settings pane with a Delete World and Restore backup Button (the Restore Backup button is still non-functional, I might comment it out for this PR and include it in another PR which I plan to add better backup logic)
- Solved an issue where clicking the "GoBack" button too soon after world creation would crash the launcher

NOTE: I commented out the code and logic for the Start Server button in the World Settings pane in case we decide later on to have it there too.

NOTE 2: World seed input does not get saved now when creating a new world for some reason - will need to investigate and fix when I try to get the world seed from the world's save files if the config file doesn't have one.