Fix remote player stuttering/teleporting when standing still or moving. (#1947)
* Fix remote player stuttering/teleporting when standing still or moving.

* Changed movement handling to use FixedUpdate

This is a static update loop that runs less frequent than .Update to reduce packet sending pressure. In addition, movement changes due to physics will be incorporated in the movement update.
Reduced the GC pressure by caching the yield returns.

* Added player location update thottle based on skipping physics updates

Previously it used a time-based throttle which could mismatch with the update loop.