SpawnOnKill and BreakableResource replication added (e.g. Coral Disks or Limestones) (#2016)
* Added SpawnOnKill behaviour

Added SpawnOnKill_Patch.cs to handle SpawnOnKill packets, still in work --- It syncs the destroy but not the spawning yet / I plan to copy the code from the original game, but if someone has the bravour to make a transpiler... then go on.

* Updated SpawnOnKill and renamed the file to match the naming convetion.
It is done ! I finished my first task ! I'm really proud 😊

* Updated SpawnOnKill_OnKill_Patch.cs to match requested changes

* Patched the game object not updating when entity is "killed".
It was dropping the original entity, so even if physic was now simulated, the other player(s) still had to knife it or kill it on their side.
Now it's working perfectly.

* Create BreakableResource_BreakIntoResources_Patch.cs

Created a patch to delete the rock when a breakableresource is broken. However, I still need to make a transpiler for function SpawnResourceFromPrefab() which is a coroutine. Wish me good luck ;-;

* Updated SpawnOnKill from a Prefix to a Transpiler. It's optimized 👌

* Removed the prefix code

* renamed Callback argument "__this" to "spawnOnKill". Updated refractors.

* Started working on SpawnResourceFromPrefab patch.

* Slight refactor of EntityDestroyedProcessor.Process to not always call expensive methods

* Added docs to Items.cs APIs

Changed Dropped() to use UWE API to get TechType from passed in GameObject if not provided.

* Cleaned up SpawnOnKill_OnKill_Patch.cs

Added to pattern match on Object.Instantiate from UnityEngine to better detect code updates that break this patch (trying to prevent silent errors).
Removed some comments that added too much information and makes code harder to understand at face-value.

* Added unit test for SpawnOnKill_OnKill_Patch.cs

* Updated SpawnResourceFromPrefab method [NOT DONE!]

* Merged BreakableResource to SpawnOnKill branches and finished BreakableResource_SpawnResourceFromPrefab_Patch, testing needed

* patched syntax error

* added class comments

* optimized SpawnResourceFromPrefab to match latest changes on .Dropped() function

* Fixed SpawnResourceFromPrefab. TEST NEEDED.

After this commit, please review the PR and tell me if there's anything else needed for the 1.8

* Added weak matching on call-like instructions when using Reflect API in pattern matching

* Removed dead code in EntityDestroyedProcessor

* Added logging when EntityDestroyedProcessor failed to find the gameobject

* corrected partially requested changes. Added a test class, used the new reflect.method for operand checks, deleted useless usings...

* corrected test class (offset was 3 instead of 2)

* Added a security to BreakIntoResources

* I failed my save, now it's secured*

* removed comment because I understood

* Disabled weak match on call opcode in IL pattern when using InstructionPattern.Call directly

* Reverted change in Items.Dropped() API to not stop on TechType.None

* edited spawnOnKill patch and added a security like in BreakIntoResources patch. Ready for merge btw

* corrected what @sunrunner told me to change

* corrected syntax error, undefined reference and else clause.

* Corrected naming and formatting. Ready for merging !

---------

Co-authored-by: Measurity <measuring.infinity@gmail.com>