Implemented Smooth Scrolling
FINALLY... it's 3:30am and I got this figured out. I have no idea how I ended up with this black magic that you see before you, but it works pretty well. Please let me know if I've committed any unforgivable sins in these changes :)

On a more serious note, I think the scroll animation looks way better than it did before, but it still feels slightly jittery. And I don't like how I had to rely on a "LastOffset" property for this to work properly, but it's the best that I could come up with seeing that "ScrollViewer.Offset" is already changed to the target scroll value when this code is being run.