Fixed TextBox style
There was a weird styling issue that I noticed where the actual "text" block area was centered within the textbox, making the cursor not change unless you hovered over it and also making selecting text a bit less convenient.

This commit fixes that, though it would've been easier to fix had it not been for the DockPanel#PART_InnerDockPanel element refusing to stay at 0 margin and instead continuing to use the TemplateBinding padding. About 2.5 hours of work had to be put in and many hoops had to be jumped through to solve this issue :(