Disable setting fields in Advanced Settings window when server is online
It will now be impossible to edit the fields within the Advanced Settings window while the server is running (which was possible before when an external server was started within the ManageServerView).

I also added a max height to DialogBoxModal so that long errors don't make the error window height go off screen like it did for me while making this commit.

Known issue: The setting fields in the Advanced Settings window will not re-enable when the external server is closed via the CMD window. This can be worked around by closing and opening the advanced settings window.