Fix server crashing not allowing server to be stopped
Added a 7 second timeout to the stop server command, where it'll then force shutdown the server if it doesn't close correctly. I decided that adding a dialog box would be hard to do and not necessary since the user can't do anything different if the server isn't forcefully shut down.

This feature can be tested by deleting a json file in a save folder.

Co-Authored-By: dartasen <10561268+dartasen@users.noreply.github.com>