Added "Copy URL" context menu for link textblocks
Users can now copy the link that "link" style textblocks have in case they don't trust clicking on it directly.

Also fixed the click event handler for link text blocks to run on both left and right click (in order to allow for the context menu to open with right click).

Additionally, I fixed a small typo that I randomly found in BlogView.