Almost finished implementing Server/Launcher IPC
The only remaining issue so far is that the logger for the imbedded server in the launcher bugs out at a very specific point during server startup (at the Autosave line) and then stops working, despite the server continuing to run normally in the background.

After an hour of investigating, I'm not sure how to fix this. Measurity pls perform your black magic :(