Fixed multiline output issue in EmbeddedServerView
Fixed an issue in EmbeddedServerView for multiline server output logs (such as that from the "summary" command) which were being treated as a single output line instead of being split per newline (fixes trailing newlines being cut and most of the scrolling issues I previously mentioned).

There is still a bug with scrolling that seems to randomly happen. It happened once to me when I used the "help" command and tried to scroll up.