Added port-in-use prompt for server startup
If a server's port is in use when the user tries to start it in the launcher, there will be a dialog box asking the user if they would like to change the port.

New issue I noticed while doing this: the ImbeddedServerView won't update the countdown for the port checking log message in the same way that the external server does. Not a release blocker, but I will look into this later when I can.