Removed SerializerMode option from the Advanced Settings view
This fixes a bug where the server disappears from the launcher servers view when users swap the serializer within the launcher. This issue also caused the server to create a fresh save when you start the server after switching this config option.

Users will now have to swap it by starting the server, running the "swapserializer (mode)" command, and then saving.