Improve ScrollViewer for ImbeddedServerView
Small improvements in ImbeddedServerView to fix a weird visual glitch with the horizontal scroll bar when output lines longer than the window width are out of view.

Currently working on fixing a major issue with smooth scrolling. Looks like it's not correctly changing the LastOffset property of the scrollviewer when it is being scrolled by something other than the scrollwheel, such as using the scroll thumbs or the autoscroll feature in ImbeddedServerView. This causes an annoying bug where the scroll animation will start at an erroneous LastOffset property, looking like a glitching scrolling animation (and sometimes breaking scrolling by freezing it in place I think, as is occasionally happening to me in ImbeddedServerView)