Improved horizontal scroll in ImbeddedServerView
Because of the virtualization of ItemsControl, the horizontal scroll bar would always be changing its width as log lines came in and out of virtualization. Now, the scrollviewer will remember its maximum width and set its MinWidth property to that value, which gives the illusion of the list not being virtualized.

The only thing this doesn't change its minimum width for is when Auto Scroll is turned off and new lines of output that are longer than the min width are not yet brought into view. But in that case, once the user scrolls to that line, the min width value will update properly.