Suppressed ToggleLights default processor log spam
This prevents the debug message "Using default packet processor for: [ToggleLights: Id: 92d232bf-611d-405f-b75e-7210b6c520dd, IsOn: False] and player 2" from spamming the server console when players turn on and off the toggle lights of a seaglide/flashlight. Not sure if this is a bad idea though; please let me know if it is.