Fixed unnecessary "Ahead (speed)" message from playing
When a different player is piloting the Cyclops and quits piloting while holding down a movement key, the Cyclops may sometimes give an "Ahead (speed)" message/announcement to other passengers when it's not needed (i.e. the speed didn't change). This also sometimes happens when a different player starts piloting the cyclops after joining the game.

This fix suppresses that message.