Added Screenshots, Saves, and Logs buttons in OptionsView
Also changed the section to be for AppData instead of just for the Logs folder.

TODO: Fix the screenshots folder path to come from a public variable in NitroxUser, and rewrite the ScreenshotManager_Initialise patch to instead patch the ScreenshotManager.Combine() method and supply the custom screenshots path from NitroxUser.